feat: Move merging environment variables from own config class in server down to base config.
