+++ Server/docker/Dockerfile
@@ -155,4 +155,4 @@ RUN npm run build
 
 ARG BUILD_ARG_APP_ENV="docker"
 ENV BUILD_APP_ENV=$BUILD_ARG_APP_ENV
-CMD npm run start-unix:$BUILD_APP_ENV
+CMD npm run start-unix:"$BUILD_APP_ENV"
